+++ probe-rs/src/architecture/arm/core/armv6m.rs
@@ -3,7 +3,9 @@
 use super::{CortexMState, Dfsr, ARM_REGISTER_FILE};
 
 use crate::architecture::arm::sequences::ArmDebugSequence;
-use crate::core::{RegisterDescription, RegisterFile, RegisterKind};
+use crate::core::{
+    RegisterDataType, RegisterDescription, RegisterFile, RegisterKind, RegisterValue,
+};
 use crate::error::Error;
 use crate::memory::{valid_32_address, Memory};
 use crate::{
@@ -433,12 +435,16 @@ const PC: RegisterDescription = RegisterDescription {
     name: "PC",
     _kind: RegisterKind::PC,
     address: CoreRegisterAddress(0b0_1111),
+    _type: RegisterDataType::UnsignedInteger,
+    size_in_bits: 32,
 };
 
 const XPSR: RegisterDescription = RegisterDescription {
     name: "XPSR",
     _kind: RegisterKind::General,
     address: CoreRegisterAddress(0b1_0000),
+    _type: RegisterDataType::UnsignedInteger,
+    size_in_bits: 32,
 };
 
 /// The state of a core that can be used to persist core state across calls to multiple different cores.
@@ -539,7 +545,7 @@ impl<'probe> CoreInterface for Armv6m<'probe> {
 
         // get pc
         Ok(CoreInformation {
-            pc: pc_value.into(),
+            pc: pc_value.try_into()?,
         })
     }
 
@@ -594,7 +600,7 @@ impl<'probe> CoreInterface for Armv6m<'probe> {
 
         // get pc
         Ok(CoreInformation {
-            pc: pc_value.into(),
+            pc: pc_value.try_into()?,
         })
     }
 
@@ -613,9 +619,9 @@ impl<'probe> CoreInterface for Armv6m<'probe> {
         let _ = self.status()?;
 
         const XPSR_THUMB: u32 = 1 << 24;
-        let xpsr_value = self.read_core_reg(XPSR.address)?;
+        let xpsr_value: u32 = self.read_core_reg(XPSR.address)?.try_into()?;
         if xpsr_value & XPSR_THUMB == 0 {
-            self.write_core_reg(XPSR.address, xpsr_value | XPSR_THUMB)?;
+            self.write_core_reg(XPSR.address, (xpsr_value | XPSR_THUMB).into())?;
         }
 
         self.sequence
@@ -626,7 +632,7 @@ impl<'probe> CoreInterface for Armv6m<'probe> {
 
         // get pc
         Ok(CoreInformation {
-            pc: pc_value.into(),
+            pc: pc_value.try_into()?,
         })
     }
 
@@ -778,12 +784,13 @@ impl<'probe> CoreInterface for Armv6m<'probe> {
         Ok(CoreStatus::Running)
     }
 
-    fn read_core_reg(&mut self, address: CoreRegisterAddress) -> Result<u32, Error> {
-        super::cortex_m::read_core_reg(&mut self.memory, address)
+    fn read_core_reg(&mut self, address: CoreRegisterAddress) -> Result<RegisterValue, Error> {
+        let val = super::cortex_m::read_core_reg(&mut self.memory, address)?;
+        Ok(val.into())
     }
 
-    fn write_core_reg(&mut self, address: CoreRegisterAddress, value: u32) -> Result<()> {
-        super::cortex_m::write_core_reg(&mut self.memory, address, value)?;
+    fn write_core_reg(&mut self, address: CoreRegisterAddress, value: RegisterValue) -> Result<()> {
+        super::cortex_m::write_core_reg(&mut self.memory, address, value.try_into()?)?;
         Ok(())
     }